project/requirements.txt
# --- Core Frameworks ---
Flask==3.1.2
Flask-Session==0.8.0
gunicorn==23.0.0
python-dotenv==1.1.1

# --- Database ---
pymongo==4.15.1

# --- AI & NLP / Resume Analysis ---
numpy==1.26.4
pandas==2.3.2
scikit-learn==1.3.0
sentence-transformers==2.2.2
torch==2.2.2
transformers==4.35.0
nltk==3.8.1

# --- Mock Interview (Voice / NLP / Media) ---
opencv-python==4.11.0.86
mediapipe==0.10.21
moviepy==1.0.3

# --- Utility Libraries ---
requests==2.32.4
tqdm==4.67.1
regex==2025.9.18
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2

# --- Optional / for AI integration (if used) ---
google-generativeai==0.8.5
openai==0.27.10
